INF loof made by using size_t data type is soloved

+++ ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: jeshin <jeshin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/06 16:52:23 by jeshin            #+#    #+#             */
-/*   Updated: 2023/10/20 16:07:06 by jeshin           ###   ########.fr       */
+/*   Updated: 2023/10/21 19:14:02 by jeshin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,7 +58,7 @@ char	*ft_strtrim(char const *s1, char const *set)
 	size_t	len;
 	size_t	l_beg;
 	size_t	l_end;
-	int		i;
+	size_t		i;
 
 	if (!s1)
 		return (0);
@@ -73,7 +73,7 @@ char	*ft_strtrim(char const *s1, char const *set)
 	if (!ret)
 		return (0);
 	i = 0;
-	while (s1[l_beg] && (size_t)i < len)
+	while (s1[l_beg] && i < len)
 		ret[i++] = s1[l_beg++];
 	ret[i] = 0;
 	return (ret);
